auth/pyproject.toml
[tool.poetry]
name = "fast-api-microservice-test"
version = "0.1.0"
description = ""
authors = ["Akinshin Anatoly <a.akinshin@satvaspace.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
PyJWT = "^2.4.0"
SQLAlchemy = "^1.4.38"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^1.2.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
